parse_xml and start_testing_mock modules
