+++ runner.py
@@ -10,10 +10,11 @@ import argparse
 import signal
 from concurrent.futures import ThreadPoolExecutor
 import threading
-import xml.etree.ElementTree as ET
 import subprocess
 import os
 from run_test import run_test
+from start_testing_mock import start_testing_mock
+from parse_xml import parse_xml_result
 
 stop_event = threading.Event()
 jobs_pool = None
@@ -27,71 +28,6 @@ class Config(object):
 def signal_handler(signum, frame):
 	print(f"\nReceived signal, terminating...")
 	stop_event.set()
-
-def parse_xml_result(report_path):
-	result = {}
-	attempt_result = {}
-	test_results = []
-
-	test_suites = ET.parse(report_path).getroot()
-	test_suit = test_suites.find('testsuite')
-
-	attempt_result["isPassed"] = int(test_suit.get("errors")) == 0 and int(test_suit.get("failures")) == 0
-	attempt_result["isError"] = int(test_suit.get("errors")) != 0
-	attempt_result["errorDetails"] = "error during task execution" if attempt_result["isError"] else ""
-	attempt_result["duration"] = test_suit.get("time")
-
-	for testcase in test_suit.findall("testcase"):
-		test_result = {}
-
-		class_name = testcase.get("classname")
-		test_name = testcase.get("name")
-		name = class_name.replace("tests.", "")
-		name = name + "::" + test_name
-
-		test_result["name"] = name
-		test_result["isPassed"] = True
-		test_result["description"] = ""
-		test_result["duration"] = testcase.get("time")
-		test_result["memoryUsed"] = random.randrange(1000, 2000)
-
-		if len(testcase):
-			failure = testcase.find("failure")
-			if failure is not None:
-				test_result["isPassed"] = False
-				test_result["description"] = failure.text
-
-		test_results.append(test_result)
-
-	attempt_result["testCases"] = test_results
-
-	result["testResults"] = attempt_result
-	return result
-
-
-def start_testing_mock(config, task): ## For backend testing
-	print(f"Received task")
-	task_json = task.json()
-
-	## TODO: Execute run_test.sh
-
-	result = {}
-
-	## For testing
-	xml_path1 = "mock_server/reports/passed_report.xml"
-	xml_path2 = "mock_server/reports/failed_report.xml"
-
-	if task_json["repositoryUrl"] == "https://github.com/os-nsu/proxy-grisha.git":
-		result = parse_xml_result(xml_path1)
-	elif task_json["repositoryUrl"] == "https://github.com/os-nsu/proxy-anton.git":
-		result = parse_xml_result(xml_path2)
-
-	result["attempt"] = task_json["attempt"]
-
-	print(f"post body: {result}")
-
-	post_results(result)
-
 
 def start_testing(config, task):
 	print(f"Received task")
